neededfiles changed

+++ TP2/src/SendHandler.java
@@ -385,11 +385,11 @@ public class SendHandler implements Runnable {
 
         sendData(ip, clientHandlerPort, seq, i, data);
 
-        long stopTimer = System.currentTimeMillis();
-        long durationInMillis = stopTimer - startTimer;
-        long durationInSeconds = durationInMillis / 1000;
+        double stopTimer = System.currentTimeMillis();
+        double durationInMillis = stopTimer - startTimer;
+        double durationInSeconds = durationInMillis / 1000;
 
-        long debit = (file.length() * 8) / durationInSeconds;
+        long debit = (long) ((file.length() * 8) / durationInSeconds);
         // regist into the log file
 
         String log = file.getName() + " was sent to " + ip + " | Duration= " + durationInMillis + "ms | Debit= " + debit
